requirements.txt
discord.py>=2.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
fake-useragent>=1.4.0
python-dotenv>=1.0.0
